Add summary for non-engineers
Many of our working agreements also affect our team members who aren't
engineers and it's important that we give them access to this information
so that they know things like what we expect from an issue and why we're
learning instead of working on a ticket on a Friday afternoon. I added
a page with a summary of the most relevant information and links for
those team members so we can ensure we're all on the same page without
expecting them to sift through content they don't need to know.

Additionally, I added some links to child pages in the table of contents
lists so they're easier to find.