Stipulate minimum commit message length
This follows up on a discussion from engineering meeting. As stated,
this is a documentation on the rule so we can have something to vote
and/or amend.

I made the decision to be overly verbose on this part because we don't
have the background of the "Seven Rules of a Good Commit Message" to
rely on the why here. I didn't change the example commit message because
it already demonstrated this rule.