Git branches are known by Pivotal
This follows up on a discussion from an engineering meeting. This is a
starting point for us to debate, amend, and vote.

The longer context in here is not in line with the rest of the document,
but I found it helpful. It's possible this would need a new page if we
started doing this for every bullet, but for now it seems fine.